requirements.txt
yfinance>=0.2.61
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
lxml>=5.0.0
